+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Conditional execute properties to submission events
 - Allowed 'NOW' for date, time and DateTime defaultValues
 
 # 0.7.2 (2020-07-09)